requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.18.1
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
tifffile==2021.1.14
tqdm==4.56.0
wget==3.2
